Don't allow ContentType to be deleted if Image objects exist
The on_delete argument will be a required argument in Django 2.0;
previously, it has defaulted to CASCADE, which is undesirable for Image.
ContentType objects may be deleted after a prompt when migrating,
and we don't want to lose Images like that.

Note that on_delete is keyword-argument-only in Django 1.8;
from Django 1.9 it can be used as the second positional argument.