One year at a time on gallery.html. WIP
Before we made two db calls per thumbnail and showed all years on a
single request. Now we make one call pr thumbnail and only show one year
per request.

This last call have to be eliminated. It may be possible to go back to
show all years when it is done.

Work on #52.