Fix get_exif_date timezone issue
get_current_timezone() would return right 'Europe/Copenhagen'-timezome. But the
difference between UTC and the local time is not constant over time. dt.replace
is naive of this. It would result in it calculating an offset of +00:50
which is the correct offset from before 1894.

This also makes it DST aware.

Work on #104