Revert "Raise status 500 version of 404 with explanation, closes #111"
This reverts commit 8581972f8ad0fd71064dcf98b30b4f2810f37f20.

This did not actually solve #111 only made the page a bit prettier for a small
subset of the pages. The user was not able to do anything to solve the problem
as the /admin/gallery/album/<id>/change/ also returned 500. (You could however
solve it by deleting the whole album from /admin/gallery/album/)

In ea89fc8b2313d380656b4b5c76eb3d52de7d8906 the problem was solved by not
allowing a user to upload a corrupt picture in the first place. In the case that
there is some outstanding bug which causes a 500 on the gallery page, the
correct way to solve it would be (other than to make sure it does not occur in
the first place), to make a global Error View. See
https://docs.djangoproject.com/en/1.11/ref/views/#error-views